Hackerrank Hanging Posters challenge
Arthur needs to hang  posters on his wall. Standing on the ground he can reach upto a height of . Each poster is to be bolted at a certain height above the ground level, described by the array . Each poster also has some length, defined by the array .

To hang a poster properly, Arthur needs to hold atleast 50% of the length of the poster and poster is to be bolted at a point which is 25% from its top. 

image
Arthur wants to know what is the minimum height of the ladder he should buy, in order to hang all the wall posters. The ladder is only available in integral heights. Arthur can reach any height upto the maximum possible height.


Input Format

The first line of the input contains two space separated integers,  and .
The next line contains  space separated integers, denoting the elements of the array .
The last line contains  space separated integers, denoting the elements of the array .

Constraints

1 ≤  < 109
1 ≤  ≤ 105
1 ≤ i ≤ 109 (0 ≤ i < )
1 ≤ i ≤ 105 (0 ≤ i < )

Output Format

Output an integer, the minimum height of the ladder required. If no ladder is required, output 0

Sample Input 0

3 5
15 11 17
5 1 2
Sample Output 0

12